rename test classes to make them be executed in maven + clean up
