+++ tests/org.eclipse.n4js.hlc.integrationtests/src/org/eclipse/n4js/hlc/integrationtests/APIReplacementN4jscJarTest.java
@@ -18,13 +18,8 @@ import org.junit.Test;
 
 /**
  * Tests API / implementation replacement when running via n4jsc.
- *
- * IMPORTANT: All the tests in this class require that n4jsc.jar exist. Before executing this test, in the console,
- * change your current to the folder {@code git/n4js/tools/scripts/}. Then inside that folder, execute the
- * {@code mvn-cp-n4jsjar.sh}.
  * <p>
- * This script executes Maven locally to create n4jsc.jar and copy the n4jsc.jar to the folder
- * git/n4js/tests/org.eclipse.n4js.hlc.tests/target required by these tests.
+ * IMPORTANT: for info on how to run this test locally, see {@link AbstractN4jscJarTest}!
  */
 public class APIReplacementN4jscJarTest extends AbstractN4jscJarTest {
 
@@ -90,7 +85,8 @@ public class APIReplacementN4jscJarTest extends AbstractN4jscJarTest {
 				"--buildType", "allprojects",
 				"--projectlocations", WORKSPACE_FOLDER,
 				"--runWith", "nodejs",
-				"--run", WORKSPACE_FOLDER + "/" + "org.eclipse.n4js.client/src/A.n4js"); // note: not defining an implementation id!
+				"--run", WORKSPACE_FOLDER + "/" + "org.eclipse.n4js.client/src/A.n4js"); // note: not defining an
+																							// implementation id!
 
 		int exitCode = p.waitFor();
 		assertEquals("runner throws exception", ErrorExitCode.EXITCODE_RUNNER_STOPPED_WITH_ERROR.getExitCodeValue(),
@@ -118,7 +114,8 @@ public class APIReplacementN4jscJarTest extends AbstractN4jscJarTest {
 				"--buildType", "allprojects",
 				"--projectlocations", WORKSPACE_FOLDER,
 				"--runWith", "nodejs",
-				"--run", WORKSPACE_FOLDER + "/" + "org.eclipse.n4js.client/src/A.n4js"); // note: not defining an implementation id!
+				"--run", WORKSPACE_FOLDER + "/" + "org.eclipse.n4js.client/src/A.n4js"); // note: not defining an
+																							// implementation id!
 
 		int exitCode = p.waitFor();
