+++ tests/org.eclipse.n4js.hlc.integrationtests/src/org/eclipse/n4js/hlc/integrationtests/CompileRunTestAllPassN4jscJarTest.java
@@ -14,26 +14,27 @@ import static org.junit.Assert.assertEquals;
 
 import java.io.IOException;
 
-import org.eclipse.n4js.hlc.base.ErrorExitCode;
 import org.junit.Test;
 
 /**
  * Uses the {@code n4jsc.jar} to compile and run all tests in a given project.
- *
- * The project contains passing as well as failing tests.
+ * <p>
+ * All tests in the project pass.
+ * <p>
+ * IMPORTANT: for info on how to run this test locally, see {@link AbstractN4jscJarTest}!
  */
-public class CompileAndRunTestsWithFailuresTest extends AbstractN4jscJarTest {
+public class CompileRunTestAllPassN4jscJarTest extends AbstractN4jscJarTest {
 
 	/** */
-	public CompileAndRunTestsWithFailuresTest() {
-		super("probands/GH-975-tester-failures", true);
+	public CompileRunTestAllPassN4jscJarTest() {
+		super("probands/GH-975-tester-allpass", true);
 	}
 
 	/**
 	 * Compiles and runs all tests in a simple N4JS test project.
 	 */
 	@Test
-	public void testRunTestsWithFailures() throws IOException, InterruptedException {
+	public void testRunTestsAllPass() throws IOException, InterruptedException {
 		logFile();
 
 		final String wsRoot = WORKSPACE_FOLDER;
@@ -46,7 +47,7 @@ public class CompileAndRunTestsWithFailuresTest extends AbstractN4jscJarTest {
 
 		int exitCode = p.waitFor();
 
-		assertEquals(ErrorExitCode.EXITCODE_TESTER_STOPPED_WITH_ERROR.getExitCodeValue(), exitCode);
+		assertEquals("The process terminates successfully with exit code 0.", 0, exitCode);
 	}
 
 }
