+++ tests/org.eclipse.n4js.hlc.integrationtests/src/org/eclipse/n4js/hlc/integrationtests/CompileRunTestWithFailuresN4jscJarTest.java
@@ -14,25 +14,28 @@ import static org.junit.Assert.assertEquals;
 
 import java.io.IOException;
 
+import org.eclipse.n4js.hlc.base.ErrorExitCode;
 import org.junit.Test;
 
 /**
  * Uses the {@code n4jsc.jar} to compile and run all tests in a given project.
- *
- * All tests in the project pass.
+ * <p>
+ * The project contains passing as well as failing tests.
+ * <p>
+ * IMPORTANT: for info on how to run this test locally, see {@link AbstractN4jscJarTest}!
  */
-public class CompileAndRunTestsAllPassTest extends AbstractN4jscJarTest {
+public class CompileRunTestWithFailuresN4jscJarTest extends AbstractN4jscJarTest {
 
 	/** */
-	public CompileAndRunTestsAllPassTest() {
-		super("probands/GH-975-tester-allpass", true);
+	public CompileRunTestWithFailuresN4jscJarTest() {
+		super("probands/GH-975-tester-failures", true);
 	}
 
 	/**
 	 * Compiles and runs all tests in a simple N4JS test project.
 	 */
 	@Test
-	public void testRunTestsAllPass() throws IOException, InterruptedException {
+	public void testRunTestsWithFailures() throws IOException, InterruptedException {
 		logFile();
 
 		final String wsRoot = WORKSPACE_FOLDER;
@@ -45,7 +48,7 @@ public class CompileAndRunTestsAllPassTest extends AbstractN4jscJarTest {
 
 		int exitCode = p.waitFor();
 
-		assertEquals("The process terminates successfully with exit code 0.", 0, exitCode);
+		assertEquals(ErrorExitCode.EXITCODE_TESTER_STOPPED_WITH_ERROR.getExitCodeValue(), exitCode);
 	}
 
 }
