+++ tests/org.eclipse.n4js.hlc.integrationtests/src/org/eclipse/n4js/hlc/integrationtests/PerformanceReportN4jscJarTest.java
@@ -26,11 +26,11 @@ import org.junit.Test;
 import com.google.common.io.CharStreams;
 
 /**
- *
+ * IMPORTANT: for info on how to run this test locally, see {@link AbstractN4jscJarTest}!
  */
-public class PerformanceReportTest extends AbstractN4jscJarTest {
+public class PerformanceReportN4jscJarTest extends AbstractN4jscJarTest {
 	/** Initializes test workspace data. */
-	public PerformanceReportTest() {
+	public PerformanceReportN4jscJarTest() {
 		super("probands/GH-1062", false);
 	}
